+++ test/data/service/qvst_service_test.mocks.dart
@@ -122,17 +122,18 @@ class MockBackendApi extends _i1.Mock implements _i4.BackendApi {
       ) as _i5.Future<_i2.HttpResponse<dynamic>>);
 
   @override
-  _i5.Future<_i2.HttpResponse<dynamic>> getAllQvst() => (super.noSuchMethod(
+  _i5.Future<_i2.HttpResponse<dynamic>> getAllQvst(bool? includeNoLongerUsed) =>
+      (super.noSuchMethod(
         Invocation.method(
           #getAllQvst,
-          [],
+          [includeNoLongerUsed],
         ),
         returnValue: _i5.Future<_i2.HttpResponse<dynamic>>.value(
             _FakeHttpResponse_0<dynamic>(
           this,
           Invocation.method(
             #getAllQvst,
-            [],
+            [includeNoLongerUsed],
           ),
         )),
       ) as _i5.Future<_i2.HttpResponse<dynamic>>);
@@ -173,18 +174,26 @@ class MockBackendApi extends _i1.Mock implements _i4.BackendApi {
 
   @override
   _i5.Future<_i2.HttpResponse<dynamic>> getAllQvstQuestionsByThemeId(
-          String? id) =>
+    String? id,
+    bool? includeNoLongerUsed,
+  ) =>
       (super.noSuchMethod(
         Invocation.method(
           #getAllQvstQuestionsByThemeId,
-          [id],
+          [
+            id,
+            includeNoLongerUsed,
+          ],
         ),
         returnValue: _i5.Future<_i2.HttpResponse<dynamic>>.value(
             _FakeHttpResponse_0<dynamic>(
           this,
           Invocation.method(
             #getAllQvstQuestionsByThemeId,
-            [id],
+            [
+              id,
+              includeNoLongerUsed,
+            ],
           ),
         )),
       ) as _i5.Future<_i2.HttpResponse<dynamic>>);
@@ -236,6 +245,24 @@ class MockBackendApi extends _i1.Mock implements _i4.BackendApi {
           Invocation.method(
             #getQvstAnswersRepo,
             [],
+          ),
+        )),
+      ) as _i5.Future<_i2.HttpResponse<dynamic>>);
+
+  @override
+  _i5.Future<_i2.HttpResponse<dynamic>> addQvstAnswersRepo(
+          Map<String, dynamic>? body) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #addQvstAnswersRepo,
+          [body],
+        ),
+        returnValue: _i5.Future<_i2.HttpResponse<dynamic>>.value(
+            _FakeHttpResponse_0<dynamic>(
+          this,
+          Invocation.method(
+            #addQvstAnswersRepo,
+            [body],
           ),
         )),
       ) as _i5.Future<_i2.HttpResponse<dynamic>>);
@@ -341,6 +368,23 @@ class MockBackendApi extends _i1.Mock implements _i4.BackendApi {
           Invocation.method(
             #importQvstFile,
             [body],
+          ),
+        )),
+      ) as _i5.Future<_i2.HttpResponse<dynamic>>);
+
+  @override
+  _i5.Future<_i2.HttpResponse<dynamic>> exportQvstQuestions() =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #exportQvstQuestions,
+          [],
+        ),
+        returnValue: _i5.Future<_i2.HttpResponse<dynamic>>.value(
+            _FakeHttpResponse_0<dynamic>(
+          this,
+          Invocation.method(
+            #exportQvstQuestions,
+            [],
           ),
         )),
       ) as _i5.Future<_i2.HttpResponse<dynamic>>);
@@ -858,6 +902,22 @@ class MockBackendApiBase extends _i1.Mock implements _i7.BackendApiBase {
               campaignId,
               token,
             ],
+          ),
+        )),
+      ) as _i5.Future<_i3.Response>);
+
+  @override
+  _i5.Future<_i3.Response> exportQvstQuestionsCsv(String? token) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #exportQvstQuestionsCsv,
+          [token],
+        ),
+        returnValue: _i5.Future<_i3.Response>.value(_FakeResponse_2(
+          this,
+          Invocation.method(
+            #exportQvstQuestionsCsv,
+            [token],
           ),
         )),
       ) as _i5.Future<_i3.Response>);